pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "terms-analyzer"
version = "0.1.0"
description = "A tool to analyze and simplify terms and conditions"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "flask>=2.0.0",
    "requests>=2.25.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0"
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
